Fix warning signed unsigned comparison

+++ core/src/OpenMP/Kokkos_OpenMPexec.cpp
@@ -322,7 +322,7 @@ void OpenMP::initialize( unsigned thread_count ,
   }
 
   // Check for over-subscription
-  if( Impl::mpi_ranks_per_node() * thread_count > Impl::processors_per_node() ) {
+  if( Impl::mpi_ranks_per_node() * long(thread_count) > Impl::processors_per_node() ) {
     std::cout << "Kokkos::OpenMP::initialize WARNING: You are likely oversubscribing your CPU cores." << std::endl;
     std::cout << "                                    Detected: " << Impl::processors_per_node() << " cores per node." << std::endl;
     std::cout << "                                    Detected: " << Impl::mpi_ranks_per_node() << " MPI_ranks per node." << std::endl;
