+++ core/src/Threads/Kokkos_ThreadsExec.cpp
@@ -747,7 +747,7 @@ void ThreadsExec::initialize( unsigned thread_count ,
   }
 
   // Check for over-subscription
-  if( Impl::mpi_ranks_per_node() * thread_count > Impl::processors_per_node() ) {
+  if( Impl::mpi_ranks_per_node() * long(thread_count) > Impl::processors_per_node() ) {
     std::cout << "Kokkos::Threads::initialize WARNING: You are likely oversubscribing your CPU cores." << std::endl;
     std::cout << "                                    Detected: " << Impl::processors_per_node() << " cores per node." << std::endl;
     std::cout << "                                    Detected: " << Impl::mpi_ranks_per_node() << " MPI_ranks per node." << std::endl;
